+++ api/app/Http/Controllers/Api/AuthController.php
@@ -3,9 +3,6 @@
 namespace App\Http\Controllers\Api;
 
 use App\User;
-use Carbon\Carbon;
-use App\PasswordReset;
-use Webpatser\Uuid\Uuid;
 use Illuminate\Http\Request;
 use Tymon\JWTAuth\Facades\JWTAuth;
 use App\Http\Controllers\Controller;
@@ -15,7 +12,7 @@ use App\Http\Requests\Api\NewPasswordModel;
 use App\Http\Requests\Api\PasswordResetModel;
 use App\Http\Requests\Api\UserRegistrationModel;
 
-class AuthController extends MailController
+class AuthController extends Controller
 {
     /**
      * Authenticate the user and create a token.
@@ -106,75 +103,5 @@ class AuthController extends MailController
         // TODO: Send email
         // TODO: Only send reset password mail once an hour
         // TODO: Handle email response and reset email
-    }
-
-    /**
-     * Send a mail for resetting the password.
-     *
-     * @param \App\Http\Requests\Api\PasswordResetModel $request
-     *
-     * @return mixed
-     */
-    public function sendResetMail(PasswordResetModel $request)
-    {
-        $userEmail = $request->email;
-        $user = User::where('email', $userEmail)->first();
-        if ($user && ! $this->isSpamming($user, config('app.password_reset_minutes'))) {
-            // User exists and had no request < app.password_reset_minutes
-            $token = Uuid::generate(4) . '-' . str_random(40);
-            $url = url('/reset/'.$token);
-            $request->request->add(['url' => $url]);
-            $request->request->add(['name' => $user->name]);
-            $this->sendPasswordResetMail($request);
-            // Store in database
-            $data = [
-                'user_id' => $user->id,
-                'token' => $token,
-                'created_at' => Carbon::now(),
-            ];
-            PasswordReset::create($data);
-        }
-    }
-
-    private function isSpamming(User $user, $minutes)
-    {
-        if ($user) {
-            $last_password_request = PasswordReset::where('user_id', $user->id)->orderBy('created_at', 'desc')->first();
-            if (! $last_password_request) {
-                // No request were stored yet
-                return false;
-            }
-            $last_password_request_time = $last_password_request->created_at;
-            return $this->isInsideInterval($last_password_request_time, $minutes);
-        }
-
-        return false;
-    }
-
-    private function isInsideInterval($lastTime, $minutes)
-    {
-        $now = Carbon::now();
-        $diff = $now->diffInMinutes(Carbon::parse($lastTime));
-        if ($diff <= $minutes) {
-            return true;
-        }
-
-        return false;
-    }
-
-    /**
-     * Store the new password in the database if token is valid and time < app.password_reset_minutes.
-     *
-     * @param \App\Http\Requests\Api\NewPasswordModel $request
-     * @param string $token
-     */
-    public function resetPassword(NewPasswordModel $request, $token)
-    {
-        $passwordReset = PasswordReset::where('token', $token)->first();
-        if ($passwordReset && $this->isInsideInterval($passwordReset->created_at, config('app.password_reset_minutes'))) {
-            $user = $passwordReset->user()->first();
-            $user->password = bcrypt($request->password);
-            $user->save();
-        }
     }
 }
