+++ api/routes/api.php
@@ -26,8 +26,9 @@ Route::group(['namespace' => 'Api'], function () {
 
         Route::post('register', 'AuthController@register');
         Route::post('refresh', 'AuthController@refresh');
-        Route::post('reset', 'AuthController@sendResetMail');
-        Route::post('reset/{token}', 'AuthController@resetPassword');
+
+        Route::post('reset', 'PasswordResetController@sendResetMail');
+        Route::post('reset/{token}', 'PasswordResetController@resetPassword');
     });
 
     // Authenticated url's for Installation devices
